Extract CMECurrencyExchangeCalendar into its own file
