+++ tests/test_exchange_calendar_cme_globex_fx.py
@@ -3,7 +3,7 @@ import datetime as dt
 import pandas as pd
 import pytz
 
-from pandas_market_calendars.exchange_calendar_cme import CMECurrencyExchangeCalendar
+from pandas_market_calendars.exchange_calendar_cme_globex_fx import CMECurrencyExchangeCalendar
 
 
 def test_time_zone():
